[HPRO-749] Migrate Participant Review to Symfony
Migrates the Participant Review component to Symfony.

[HPRO-749]
