+++ src/Pmi/Review/Review.php
@@ -2,10 +2,6 @@
 namespace Pmi\Review;
 
 use Pmi\Order\Order;
-use Pmi\Evaluation\Evaluation;
-use Symfony\Component\Form\Extension\Core\Type;
-use Symfony\Component\Form\Extension\Core\Type\FormType;
-use Symfony\Component\Validator\Constraints;
 
 class Review
 {
@@ -20,140 +16,6 @@ class Review
         'processed_ts' => 'Processed',
         'finalized_ts' => 'Finalized'
     ];
-
-    protected static $measurementsStatus = [
-        'created_ts' => 'Created',
-        'finalized_ts' => 'Finalized'
-    ];
-
-    protected static $emptyParticipant = [
-        'orders' => null,
-        'ordersCount' => 0,
-        'physicalMeasurements' => null,
-        'physicalMeasurementsCount' => 0,
-    ];
-
-    protected function getTodayRows($startTime, $endTime, $site)
-    {
-        $ordersQuery = 'SELECT o.participant_id, \'order\' as type, o.id, null as parent_id, o.order_id, o.rdr_id, o.biobank_id, o.created_ts, o.collected_ts, o.processed_ts, o.finalized_ts, o.finalized_samples, o.biobank_finalized, ' .
-            'greatest(coalesce(o.created_ts, 0), coalesce(o.collected_ts, 0), coalesce(o.processed_ts, 0), coalesce(o.finalized_ts, 0), coalesce(oh.created_ts, 0)) AS latest_ts, ' .
-            'oh.type as h_type, ' .
-            'u.email as created_by ' .
-            'FROM orders o ' .
-            'LEFT JOIN orders_history oh ' .
-            'ON o.history_id = oh.id ' .
-            'LEFT JOIN users u ' .
-            'ON o.user_id = u.id WHERE ' .
-            '(o.site = :site OR o.collected_site = :site OR o.processed_site = :site OR o.finalized_site = :site) ' .
-            'AND ((o.created_ts >= :startTime AND o.created_ts < :endTime) ' .
-            'OR (o.collected_ts >= :startTime AND o.collected_ts < :endTime) ' .
-            'OR (o.processed_ts >= :startTime AND o.processed_ts < :endTime) ' .
-            'OR (o.finalized_ts >= :startTime AND o.finalized_ts < :endTime) ' .
-            'OR (oh.created_ts >= :startTime AND oh.created_ts < :endTime)) ';
-        $measurementsQuery = 'SELECT e.participant_id, \'evaluation\' as type, e.id, e.parent_id, null, e.rdr_id, null, e.created_ts, null, null, e.finalized_ts, null, null, ' .
-            'greatest(coalesce(e.created_ts, 0), coalesce(e.finalized_ts, 0), coalesce(eh.created_ts, 0)) as latest_ts, ' .
-            'eh.type as h_type, ' .
-            'null ' .
-            'FROM evaluations e ' .
-            'LEFT JOIN evaluations_history eh ' .
-            'ON e.history_id = eh.id WHERE ' .
-            'e.id NOT IN (SELECT parent_id FROM evaluations WHERE parent_id IS NOT NULL) ' .
-            'AND (e.site = :site OR e.finalized_site = :site) ' .
-            'AND ((e.created_ts >= :startTime AND e.created_ts < :endTime) ' .
-            'OR (e.finalized_ts >= :startTime AND e.finalized_ts < :endTime) ' .
-            'OR (eh.created_ts >= :startTime AND eh.created_ts < :endTime)) ';
-        $query = "($ordersQuery) UNION ($measurementsQuery) ORDER BY latest_ts DESC";
-
-        return $this->db->fetchAll($query, [
-            'startTime' => $startTime,
-            'endTime' => $endTime,
-            'site' => $site
-        ]);
-    }
-
-    public function getTodayParticipants($startTime, $endTime, $site = null)
-    {
-        $participants = [];
-        foreach ($this->getTodayRows($startTime, $endTime, $site) as $row) {
-            $participantId = $row['participant_id'];
-            if (!array_key_exists($participantId, $participants)) {
-                $participants[$participantId] = self::$emptyParticipant;
-            }
-            switch ($row['type']) {
-                case 'order':
-                    // Get order status
-                    foreach (self::$orderStatus as $field => $status) {
-                        if ($row[$field]) {
-                            $row['status'] = self::getOrderStatus($row, $status);
-                        }
-                    }
-                    // Get number of finalized samples
-                    if ($row['finalized_samples'] && ($samples = json_decode($row['finalized_samples'])) && is_array($samples)) {
-                        $row['finalizedSamplesCount'] = count($samples);
-                    }
-                    $participants[$participantId]['orders'][] = $row;
-                    $participants[$participantId]['ordersCount']++;
-                    break;
-                case 'evaluation':
-                    // Get physical measurements status
-                    foreach (self::$measurementsStatus as $field => $status) {
-                        if ($row[$field]) {
-                            $row['status'] = $this->getEvaluationStatus($row, $status);
-                        }
-                    }
-                    $participants[$participantId]['physicalMeasurements'][] = $row;
-                    $participants[$participantId]['physicalMeasurementsCount']++;
-                    break;
-            }
-        }
-
-        return $participants;
-    }
-
-    public function getTodayFilterForm($formFactory, $timeZone)
-    {
-        $formBuilder = $formFactory->createBuilder(FormType::class);
-        $constraintDate = new \DateTime('today', new \DateTimeZone($timeZone));
-        $formBuilder
-            ->add('start_date', Type\DateTimeType::class, [
-                'required' => true,
-                'label' => 'Start Date',
-                'widget' => 'single_text',
-                'format' => 'M/d/yyyy',
-                'model_timezone' => $timeZone,
-                'view_timezone' => $timeZone,
-                'constraints' => [
-                    new Constraints\DateTime(),
-                    new Constraints\LessThanOrEqual([
-                        'value' => $constraintDate,
-                        'message' => 'Date cannot be in the future'
-                    ])
-                ]
-            ])
-            ->add('end_date', Type\DateTimeType::class, [
-                'required' => false,
-                'label' => 'End Date',
-                'widget' => 'single_text',
-                'format' => 'M/d/yyyy',
-                'model_timezone' => $timeZone,
-                'view_timezone' => $timeZone,
-                'constraints' => [
-                    new Constraints\DateTime(),
-                    new Constraints\GreaterThanOrEqual([
-                        'propertyPath' => 'parent.all[start_date].data',
-                        'message' => 'End date should be greater than start date'
-                    ]),
-                    new Constraints\LessThanOrEqual([
-                        'value' => $constraintDate,
-                        'message' => 'Date cannot be in the future'
-                    ])
-                ]
-            ])
-            ->add('Submit', Type\SubmitType::class, [
-                'attr' => ['class' => 'btn btn-primary'],
-            ]);
-        return $formBuilder->getForm();
-    }
 
     protected function getTodayOrderRows($today)
     {
@@ -211,20 +73,6 @@ class Review
             $status = 'Processed';
         } elseif (!empty($row['finalized_ts']) && $row['biobank_finalized']) {
             $status = 'Biobank Finalized';
-        }
-        return $status;
-    }
-
-    public function getEvaluationStatus($row, $status)
-    {
-        if ($row['h_type'] === Evaluation::EVALUATION_CANCEL) {
-            $status = 'Cancelled';
-        } elseif (!empty($row['parent_id']) && empty($row['rdr_id'])){
-            $status = 'Unlocked';
-        } elseif (!empty($row['parent_id']) && !empty($row['rdr_id'])) {
-            $status = 'Edited & Finalized';
-        } elseif (!empty($row['finalized_ts']) && empty($row['rdr_id'])) {
-            $status = 'Created';
         }
         return $status;
     }
