+++ symfony/templates/review/_macros.html.twig
@@ -1,0 +1,23 @@
+{% macro displayDate(entity, dateType) %}
+  {% if entity %}
+    {% if entity[dateType] %}
+      <span {% if entity['type'] is defined %}data-date-type="{{ entity['type'] ~ '_' ~ dateType }}"{% endif %}>
+        {% if 'now'|date('Y-m-d', app.user.timezone) == entity[dateType]|date('Y-m-d', app.user.timezone) %}
+        {# previously displayed only time if date is today. leaving this here in case we want to display something different in the future #}
+        {{ entity[dateType]|date('n/j/Y g:ia', app.user.timezone) }}
+        {% else %}
+        {{ entity[dateType]|date('n/j/Y g:ia', app.user.timezone) }}
+        {% endif %}
+      </span>
+    {% else %}
+      <i class="fa fa-times text-danger" aria-hidden="true"></i>
+    {% endif %}
+  {% endif %}
+{% endmacro %}
+
+{% macro displayLink(entity, field, participantId, idType) %}
+  {% if entity %}
+    {% set idType = entity['type'] == 'order' ? 'orderId' : 'evalId' %}
+      <a href="{{ path(entity['type'], { participantId: participantId, (idType): entity['id'] }) }}">{{ entity[field] }}</a>
+  {% endif %}
+{% endmacro %}
