+++ symfony/templates/review/modified_measurements.html.twig
@@ -1,0 +1,57 @@
+{% extends 'base.html.twig' %}
+{% block title %}Modified Measurements - Participant Review - {% endblock %}
+{% block bodycontainer %}container-fluid{% endblock %}
+{% import 'review/_macros.html.twig' as today %}
+
+{% block body %}
+{% include 'review/_page_heading.html.twig' %}
+
+<div class="container">
+  <table class="table table-striped table-bordered">
+    <thead>
+      <tr>
+        <th>Participant ID</th>
+        <th>Created</th>
+        <th>Finalized</th>
+        <th>Modified</th>
+        <th>Status</th>
+      </tr>
+    </thead>
+    <tbody>
+      {% for measurement in measurements %}
+      <tr>
+        <td>
+          <a href="{{ path('evaluation', { participantId: measurement.participant_id, evalId: measurement.id }) }}">{{ measurement.participant_id }}</a>
+        </td>
+        <td>
+          {{ today.displayDate(measurement, 'created_ts') }}
+        </td>
+        <td>
+          {% if measurement.rdr_id is not empty %}
+          {{ today.displayDate(measurement, 'finalized_ts') }}
+          {% else %}
+          <i class="fa fa-times text-danger" aria-hidden="true"></i>
+          {% endif %}
+        </td>
+        <td>
+          {{ today.displayDate(measurement, 'modified_ts') }}
+        </td>
+        <td align="center">
+          {% if measurement.eh_type == 'cancel' %}
+          <span class="label label-danger">Cancelled</span>
+          {% elseif measurement.finalized_ts is empty or (measurement.finalized_ts is not empty and measurement.rdr_id is empty) %}
+          <span class="label label-warning">Edited &amp; Unfinalized</span>
+          {% elseif measurement.rdr_id is not empty %}
+          <span class="label label-success">Edited &amp; Finalized</span>
+          {% endif %}
+        </td>
+      </tr>
+      {% else %}
+      <tr>
+        <td colspan="99" align="center">No records found</td>
+      </tr>
+      {% endfor %}
+    </tbody>
+  </table>
+</div>
+{% endblock %}
