+++ symfony/templates/review/today.html.twig
@@ -1,0 +1,143 @@
+{% extends 'base.html.twig' %}
+{% block title %}Today - Participant Review - {% endblock %}
+{% block bodycontainer %}container-fluid{% endblock %}
+{% set missingName = '' %}
+{% import 'review/_macros.html.twig' as today %}
+
+{% block body %}
+{% include 'review/_page_heading.html.twig' %}
+
+<div class="well today-filter-well" data-title="Filter">
+  {{ form_start(todayFilterForm, { attr: { class: 'form-inline', autoComplete: 'off' } }) }}
+  {{ form_errors(todayFilterForm) }}
+  <div class="today-filter" align="center">
+      {{ form_rest(todayFilterForm) }}
+      <div class="form-group">
+          <a href="{{ path('review_today') }}" class="btn btn-default">Reset</a>
+      </div>
+  </div>
+  {{ form_end(todayFilterForm) }}
+</div>
+<div class="text-center text-warning">
+    <strong>{{ displayMessage }}</strong>
+</div>
+<table class="table table-bordered table-small" id="table-today" data-name-lookup-url="{{ path('participant_review_name_lookup', { csrf_token: csrf_token('review'), id: '' }) }}" data-missing-name="{{ missingName }}">
+  <thead>
+    <tr>
+      <th rowspan="2">Participant ID</th>
+      <th rowspan="2">Name</th>
+      <th colspan="3" class="text-center td-left-border" style="background-color: #f9f9f9">Physical Measurement</th>
+      <th colspan="9" class="text-center td-left-border" style="background-color: #f9f9f9">Biobank Order</th>
+    </tr>
+    <tr>
+      <th class="td-left-border">Status</th>
+      <th>Created <span class="badge" id="evaluation_created_ts"></span></th>
+      <th>Finalized <span class="badge badge-success" id="evaluation_finalized_ts"></span></th>
+      <th class="td-left-border">Status</th>
+      <th>Biobank ID</th>
+      <th>Order ID</th>
+      <th>Created By</th>
+      <th>Created <span class="badge" id="order_created_ts"></span></th>
+      <th>Collected <span class="badge badge-warning" id="order_collected_ts"></span></th>
+      <th>Processed <span class="badge badge-info" id="order_processed_ts"></span></th>
+      <th>Finalized <span class="badge badge-success" id="order_finalized_ts"></span></th>
+      <th>Finalized Samples</th>
+    </tr>
+  </thead>
+  <tbody>
+    {% for id, participant in participants %}
+      {% set maxCount = max(participant.ordersCount, participant.physicalMeasurementsCount) %}
+      {% set rowSpan = 'rowspan =' ~ maxCount %}
+      {% for i in 0..maxCount - 1 %}
+      <tr>
+        {% if i == 0 %}
+        <td {{ rowSpan }} class="align-middle"><a href="{{ path('participant', { id: id }) }}">{{ id }}</a></td>
+        {% if participant.participant is defined %}
+        <td {{ rowSpan }} class="align-middle">
+          {% if participant.participant %}
+          <a href="{{ path('participant', { id: id }) }}">{{ participant.participant.lastName|default('') }}, {{ participant.participant.firstName|default('') }}</a>
+          {% else %}
+          {{ missingName }}
+          {% endif %}
+        </td>
+        {% else %}
+        <td {{ rowSpan }} data-participant-id="{{ id }}" data-href="{{ path('participant', { id: id }) }}" class="load-name align-middle">
+          <i class="fa fa-spinner fa-spin" aria-hidden="true"></i>
+          <span class="sr-only">Loading...</span>
+        </td>
+        {% endif %}
+        {% endif %}
+        {% set physicalMeasurement = participant.physicalMeasurements[i] is defined ? participant.physicalMeasurements[i] : '' %}
+        <td class="td-left-border">
+          {{ today.displayLink(physicalMeasurement, 'status', id) }}
+        </td>
+        <td>
+          {{ today.displayDate(physicalMeasurement, 'created_ts') }}
+        </td>
+        <td>
+          {% if physicalMeasurement %}
+          {% if physicalMeasurement.h_type != 'unlock' and not (physicalMeasurement.finalized_ts and physicalMeasurement.rdr_id is empty) %}
+          {{ today.displayDate(physicalMeasurement, 'finalized_ts') }}
+          {% else %}
+          <i class="fa fa-times text-danger" aria-hidden="true"></i>
+          {% endif %}
+          {% endif %}
+        </td>
+        {% set order = participant.orders[i] is defined ? participant.orders[i] : '' %}
+        <td class="td-left-border">
+          {{ today.displayLink(order, 'status', id) }}
+        </td>
+        <td>
+          {{ today.displayLink(order, 'biobank_id', id) }}
+        </td>
+        <td>
+          {{ today.displayLink(order, 'order_id', id) }}
+        </td>
+        <td>
+          {% if order %}
+          {{ order.created_by }}
+          {% endif %}
+        </td>
+        <td>
+          {{ today.displayDate(order, 'created_ts') }}
+        </td>
+        <td>
+          {{ today.displayDate(order, 'collected_ts') }}
+        </td>
+        <td>
+          {{ today.displayDate(order, 'processed_ts') }}
+        </td>
+        <td>
+          {% if order %}
+          {% set showOrderFinalizedTs = order.h_type != 'unlock' and not (order.finalized_ts and order.rdr_id is empty) %}
+          {% if showOrderFinalizedTs %}
+          {{ today.displayDate(order, 'finalized_ts') }}
+          {% else %}
+          <i class="fa fa-times text-danger" aria-hidden="true"></i>
+          {% endif %}
+          {% endif %}
+        </td>
+        <td>
+          {% if order %}
+          {% if showOrderFinalizedTs and order.finalizedSamplesCount is defined %}
+          {{ order.finalizedSamplesCount }}
+          {% else %}
+          <i class="fa fa-times text-danger" aria-hidden="true"></i>
+          {% endif %}
+          {% endif %}
+        </td>
+      </tr>
+      {% endfor %}
+    {% else %}
+    <tr>
+      <td colspan="99" align="center">No records found</td>
+    </tr>
+    {% endfor %}
+  </tbody>
+</table>
+
+{% endblock %}
+
+{% block pagejs %}
+    {{ encore_entry_script_tags('today') }}
+{% endblock %}
