+++ symfony/templates/review/unfinalized_measurements.html.twig
@@ -1,0 +1,51 @@
+{% extends 'base.html.twig' %}
+{% block title %}Unfinalized Measurements - Participant Review - {% endblock %}
+{% block bodycontainer %}container-fluid{% endblock %}
+{% import 'review/_macros.html.twig' as today %}
+
+{% block body %}
+{% include 'review/_page_heading.html.twig' %}
+
+<div class="container">
+  <table class="table table-striped table-bordered">
+    <thead>
+      <tr>
+        <th>Participant ID</th>
+        <th>Created</th>
+        <th>Finalized</th>
+        <th></th>
+      </tr>
+    </thead>
+    <tbody>
+      {% for measurement in measurements %}
+      <tr>
+        <td>
+          <a href="{{ path('evaluation', { participantId: measurement.participant_id, evalId: measurement.id }) }}">{{ measurement.participant_id }}</a>
+        </td>
+        <td>
+          {{ today.displayDate(measurement, 'created_ts') }}
+        </td>
+        <td>
+          {% if measurement.parent_id is empty %}
+          {{ today.displayDate(measurement, 'finalized_ts') }}
+          {% else %}
+          <i class="fa fa-times text-danger" aria-hidden="true"></i>
+          <span class="badge">Unlocked</span>
+          {% endif %}
+        </td>
+        <td>
+          {% if measurement.parent_id is empty %}
+          <a href="{{ path('evaluation_modify', { participantId: measurement.participant_id, evalId: measurement.id, type: 'cancel' }) }}" class="btn btn-xs btn-danger">Cancel Physical Measurements</a>
+          {% endif %}
+        </td>
+      </tr>
+      {% else %}
+      <tr>
+        <td colspan="99" align="center">No records found</td>
+      </tr>
+      {% endfor %}
+
+    </tbody>
+  </table>
+</div>
+{% endblock %}
