+++ symfony/templates/review/unfinalized_orders.html.twig
@@ -1,0 +1,62 @@
+{% extends 'base.html.twig' %}
+{% block title %}Unfinalized Orders - Participant Review - {% endblock %}
+{% block bodycontainer %}container-fluid{% endblock %}
+{% import 'review/_macros.html.twig' as today %}
+
+{% block body %}
+{% include 'review/_page_heading.html.twig' %}
+
+<div class="container">
+  <table class="table table-striped table-bordered">
+    <thead>
+      <tr>
+        <th>Order ID</th>
+        <th>Status</th>
+        <th>Created</th>
+        <th>Collected</th>
+        <th>Processed</th>
+        <th>Finalized</th>
+        <th></th>
+      </tr>
+    </thead>
+    <tbody>
+      {% for order in orders %}
+      <tr>
+        <td>
+          <a href="{{ path('order', { participantId: order.participant_id, orderId: order.id }) }}">{{ order.order_id }}</a>
+        </td>
+        <td>
+          <nobr>{{ order.orderStatus }}</nobr>
+        </td>
+        <td>
+          {{ today.displayDate(order, 'created_ts') }}
+        </td>
+        <td>
+          {{ today.displayDate(order, 'collected_ts') }}
+        </td>
+        <td>
+          {{ today.displayDate(order, 'processed_ts') }}
+        </td>
+        <td>
+          {% if order.h_type != 'unlock' %}
+          {{ today.displayDate(order, 'finalized_ts') }}
+          {% else %}
+          <i class="fa fa-times text-danger" aria-hidden="true"></i>
+          <span class="badge">Unlocked</span>
+          {% endif %}
+        </td>
+        <td>
+          {% if order.h_type != 'unlock' %}
+          <a href="{{ path('order_modify', { participantId: order.participant_id, orderId: order.id, type: 'cancel' }) }}" class="btn btn-xs btn-danger">Cancel Order</a>
+          {% endif %}
+        </td>
+      </tr>
+      {% else %}
+      <tr>
+        <td colspan="99" align="center">No records found</td>
+      </tr>
+      {% endfor %}
+    </tbody>
+  </table>
+</div>
+{% endblock %}
