+++ tests/Pmi/AbstractWebTestCase.php
@@ -86,7 +86,6 @@ abstract class AbstractWebTestCase extends TestCase
         $app->mount('/', new Controller\SymfonyMigrationController());
         $app->mount('/dashboard', new Controller\DashboardController());
         $app->mount('/workqueue', new Controller\WorkQueueController());
-        $app->mount('/review', new Controller\ReviewController());
 
         return $app;
     }
