--- views/review/measurements-recent-modify.html.twig
@@ -1,71 +1,0 @@
-{% extends 'base.html.twig' %}
-{% block title %}Measurements - Participant Review - {% endblock %}
-{% block bodycontainer %}container-fluid{% endblock %}
-
-{% macro displayDate(order, dateType) %}
-    {% if order[dateType] %}
-        {{ order[dateType]|date('n/j/Y g:ia', app.userTimezone) }}
-    {% else %}
-        <i class="fa fa-times text-danger" aria-hidden="true"></i>
-    {% endif %}
-{% endmacro %}
-
-{% import _self as today %}
-
-{% block body %}
-    <div class="page-header">
-        <h2><i class="fa fa-list" aria-hidden="true"></i> Participant Review</h2>
-    </div>
-    <ul class="nav nav-tabs">
-        <li role="presentation"><a href="{{ path('review_today') }}">Today's Participants</a></li>
-        <li role="presentation"><a href="{{ path('review_orders') }}">Unfinalized Biobank Orders</a></li>
-        <li role="presentation"><a href="{{ path('review_measurements') }}">Unfinalized Physical Measurements</a></li>
-        <li role="presentation"><a href="{{ path('review_ordersRecentModify') }}">Recently Modified Biobank Orders</a></li>
-        <li role="presentation" class="active"><a href="{{ path('review_measurementsRecentModify') }}">Recently Modified Physical Measurements</a></li>
-    </ul>
-    <br />
-    <div class="container">
-        <table class="table table-striped table-bordered">
-            <thead>
-            <tr>
-                <th>Participant ID</th>
-                <th>Created</th>
-                <th>Finalized</th>
-                <th>Modified</th>
-                <th>Status</th>
-            </tr>
-            </thead>
-            <tbody>
-            {% for measurement in measurements %}
-                <tr>
-                    <td>
-                        <a href="{{ path('evaluation', { participantId: measurement.participant_id, evalId: measurement.id }) }}">{{ measurement.participant_id }}</a>
-                    </td>
-                    <td>
-                        {{ today.displayDate(measurement, 'created_ts') }}
-                    </td>
-                    <td>
-                        {% if measurement.rdr_id is not empty %}
-                            {{ today.displayDate(measurement, 'finalized_ts') }}
-                        {% else %}
-                            <i class="fa fa-times text-danger" aria-hidden="true"></i>
-                        {% endif %}
-                    </td>
-                    <td>
-                        {{ today.displayDate(measurement, 'modified_ts') }}
-                    </td>
-                    <td align="center">
-                        {% if measurement.eh_type == 'cancel' %}
-                            <span class="label label-danger">Cancelled</span>
-                        {% elseif measurement.finalized_ts is empty or (measurement.finalized_ts is not empty and measurement.rdr_id is empty) %}
-                            <span class="label label-warning">Edited &amp; Unfinalized</span>
-                        {% elseif measurement.rdr_id is not empty %}
-                            <span class="label label-success">Edited &amp; Finalized</span>
-                        {% endif %}
-                    </td>
-                </tr>
-            {% endfor %}
-            </tbody>
-        </table>
-    </div>
-{% endblock %}
