--- views/review/measurements.html.twig
@@ -1,64 +1,0 @@
-{% extends 'base.html.twig' %}
-{% block title %}Measurements - Participant Review - {% endblock %}
-{% block bodycontainer %}container-fluid{% endblock %}
-
-{% macro displayDate(order, dateType) %}
-    {% if order[dateType] %}
-        {{ order[dateType]|date('n/j/Y g:ia', app.userTimezone) }}
-    {% else %}
-        <i class="fa fa-times text-danger" aria-hidden="true"></i>
-    {% endif %}
-{% endmacro %}
-
-{% import _self as today %}
-
-{% block body %}
-    <div class="page-header">
-        <h2><i class="fa fa-list" aria-hidden="true"></i> Participant Review</h2>
-    </div>
-    <ul class="nav nav-tabs">
-        <li role="presentation"><a href="{{ path('review_today') }}">Today's Participants</a></li>
-        <li role="presentation"><a href="{{ path('review_orders') }}">Unfinalized Biobank Orders</a></li>
-        <li role="presentation" class="active"><a href="#">Unfinalized Physical Measurements</a></li>
-        <li role="presentation"><a href="{{ path('review_ordersRecentModify') }}">Recently Modified Biobank Orders</a></li>
-        <li role="presentation"><a href="{{ path('review_measurementsRecentModify') }}">Recently Modified Physical Measurements</a></li>
-    </ul>
-    <br />
-    <div class="container">
-        <table class="table table-striped table-bordered">
-            <thead>
-                <tr>
-                    <th>Participant ID</th>
-                    <th>Created</th>
-                    <th>Finalized</th>
-                    <th></th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for measurement in measurements %}
-                    <tr>
-                        <td>
-                            <a href="{{ path('evaluation', { participantId: measurement.participant_id, evalId: measurement.id }) }}">{{ measurement.participant_id }}</a>
-                        </td>
-                        <td>
-                            {{ today.displayDate(measurement, 'created_ts') }}
-                        </td>
-                        <td>
-                            {% if measurement.parent_id is empty %}
-                                {{ today.displayDate(measurement, 'finalized_ts') }}
-                            {% else %}
-                                <i class="fa fa-times text-danger" aria-hidden="true"></i>
-                                <span class="badge">Unlocked</span>
-                            {% endif %}
-                        </td>
-                        <td>
-                            {% if measurement.parent_id is empty %}
-                                <a href="{{ path('evaluationModify', { participantId: measurement.participant_id, evalId: measurement.id, type: 'cancel' }) }}" class="btn btn-xs btn-danger">Cancel Physical Measurements</a>
-                            {% endif %}
-                        </td>
-                    </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-{% endblock %}
