--- views/review/orders.html.twig
@@ -1,76 +1,0 @@
-{% extends 'base.html.twig' %}
-{% block title %}Orders - Participant Review - {% endblock %}
-{% block bodycontainer %}container-fluid{% endblock %}
-
-{% macro displayDate(order, dateType) %}
-    {% if order[dateType] %}
-        {{ order[dateType]|date('n/j/Y g:ia', app.userTimezone) }}
-    {% else %}
-        <i class="fa fa-times text-danger" aria-hidden="true"></i>
-    {% endif %}
-{% endmacro %}
-
-{% import _self as today %}
-
-{% block body %}
-    <div class="page-header">
-        <h2><i class="fa fa-list" aria-hidden="true"></i> Participant Review</h2>
-    </div>
-    <ul class="nav nav-tabs">
-        <li role="presentation"><a href="{{ path('review_today') }}">Today's Participants</a></li>
-        <li role="presentation" class="active"><a href="#">Unfinalized Biobank Orders</a></li>
-        <li role="presentation"><a href="{{ path('review_measurements') }}">Unfinalized Physical Measurements</a></li>
-        <li role="presentation"><a href="{{ path('review_ordersRecentModify') }}">Recently Modified Biobank Orders</a></li>
-        <li role="presentation"><a href="{{ path('review_measurementsRecentModify') }}">Recently Modified Physical Measurements</a></li>
-    </ul>
-    <br />
-    <div class="container">
-        <table class="table table-striped table-bordered">
-            <thead>
-                <tr>
-                    <th>Order ID</th>
-                    <th>Status</th>
-                    <th>Created</th>
-                    <th>Collected</th>
-                    <th>Processed</th>
-                    <th>Finalized</th>
-                    <th></th>
-                </tr>
-            </thead>
-            <tbody>
-                {% for order in orders %}
-                    <tr>
-                        <td>
-                            <a href="{{ path('order', { participantId: order.participant_id, orderId: order.id }) }}">{{ order.order_id }}</a>
-                        </td>
-                        <td>
-                            <nobr>{{ order.orderStatus }}</nobr>
-                        </td>
-                        <td>
-                            {{ today.displayDate(order, 'created_ts') }}
-                        </td>
-                        <td>
-                            {{ today.displayDate(order, 'collected_ts') }}
-                        </td>
-                        <td>
-                            {{ today.displayDate(order, 'processed_ts') }}
-                        </td>
-                        <td>
-                            {% if order.h_type != 'unlock' %}
-                                {{ today.displayDate(order, 'finalized_ts') }}
-                            {% else %}
-                                <i class="fa fa-times text-danger" aria-hidden="true"></i>
-                                <span class="badge">Unlocked</span>
-                            {% endif %}
-                        </td>
-                        <td>
-                            {% if order.h_type != 'unlock' %}
-                                <a href="{{ path('orderModify', { participantId: order.participant_id, orderId: order.id, type: 'cancel' }) }}" class="btn btn-xs btn-danger">Cancel Order</a>
-                            {% endif %}
-                        </td>
-                    </tr>
-                {% endfor %}
-            </tbody>
-        </table>
-    </div>
-{% endblock %}
