+++ web/assets/js/views/today.js
@@ -28,5 +28,5 @@ $(document).ready(function () {
         $('#' + dateType).html($('[data-date-type=' + dateType + ']').length);
     }
 
-    $('#form_start_date, #form_end_date').pmiDateTimePicker({format: 'MM/DD/YYYY'});
+    $('#form_start_date, #form_end_date, #review_today_filter_start_date, #review_today_filter_end_date').pmiDateTimePicker({format: 'MM/DD/YYYY'});
 });
